+++ CHANGELOG.md
@@ -10,3 +10,4 @@ CHANGELOG
 * Change bundle alias from `bo_shurik_telegram_bot` to `boshurik_telegram_bot`
 * Split `bin/console telegram:webhook` command to `bin/console telegram:webhook:set` 
 and `bin/console telegram:webhook:unset`
+* Support autoconfigure for `BoShurik\TelegramBotBundle\Telegram\Command\CommandInterface` interface